Implement/fix constant pool writing for classes, packages, modules and classloaders (#17)
* Implement/fix constant pool writing for classes, packages, modules and classloaders

* Fetch JFR type IDs from host JDK instead of hardwiring them.

* Precompute and cache type IDs

* Fold writing of packages, modules and classloaders into JfrTypeRepository

* Removed debug printlns

* Renamed count variables to poolCount

* Simplify some collections accessors

* Return int from JfrTraceIdLoadBarrier.classCount()

* Improve comment

* Improve JfrTypes enum construction

* Add previousEpoch field to all JfrSymbolRepository.getSymbolId()

* Remove redundant cast

+++ substratevm/src/com.oracle.svm.jfr/src/com/oracle/svm/jfr/traceid/JfrTraceIdLoadBarrier.java
@@ -71,7 +71,7 @@ public class JfrTraceIdLoadBarrier {
         }
     }
 
-    public static long classCount(boolean epoch) {
+    public static int classCount(boolean epoch) {
         assert VMOperation.isInProgressAtSafepoint();
         return epoch ? classCount1.get() : classCount0.get();
     }
